moves workgroupsModule to internal workgroups folder and creates workgroupModule on top level of workgroup folder
